Add origin valiation to authentication requests

+++ src/api/views/auth/logout_view.py
@@ -1,19 +1,24 @@
 from django.contrib.auth.signals import user_logged_out
 
 from rest_framework import status
-from rest_framework.views import APIView
 from rest_framework.request import Request
 from rest_framework.response import Response
 
+from src.core.utils import OriginAPIView
 
-class LogoutView(APIView):
+
+class LogoutView(OriginAPIView):
     """
     View to log out the currently logged in user.
 
     It takes in an empty POST-request with an authentication header and logs out the current user, which means that the auth_token is deleted.
     """
 
+    origins = ["web", "app"]
+
     def post(self, request: Request, format=None) -> Response:
+        if (resp := super().post(request, format)) is not None:
+            return resp
         request._auth.delete()
         user_logged_out.send(
             sender=request.user.__class__, request=request, user=request.user
